+++ src/dotnet/ReSharperPlugin.SpecflowRiderPlugin/QuickFixes/CreateMissingStep/CreateMissingStepQuickFix.cs
@@ -3,8 +3,9 @@ using JetBrains.ReSharper.Feature.Services.Intentions;
 using JetBrains.ReSharper.Feature.Services.QuickFixes;
 using JetBrains.ReSharper.Feature.Services.Resources;
 using JetBrains.Util;
+using ReSharperPlugin.SpecflowRiderPlugin.Caching.StepsDefinitions;
 using ReSharperPlugin.SpecflowRiderPlugin.Daemon.Errors;
-using ReSharperPlugin.SpecflowRiderPlugin.Utils.Steps;
+using ReSharperPlugin.SpecflowRiderPlugin.Utils;
 
 namespace ReSharperPlugin.SpecflowRiderPlugin.QuickFixes.CreateMissingStep
 {
@@ -20,9 +21,18 @@ namespace ReSharperPlugin.SpecflowRiderPlugin.QuickFixes.CreateMissingStep
 
         public IEnumerable<IntentionAction> CreateBulbItems()
         {
+            var psiServices = _error.GherkinStep.GetPsiServices();
+
             return new List<IntentionAction>
             {
-                new IntentionAction(new CreateSpecflowStepFromUsageAction(_error.GherkinStep.GetStepReference(), _error.GherkinStep.GetPsiServices().GetComponent<IStepDefinitionBuilder>()), BulbThemedIcons.RedBulb.Id, IntentionsAnchors.QuickFixesAnchor)
+                new IntentionAction(new CreateSpecflowStepFromUsageAction(
+                    _error.GherkinStep.GetStepReference(),
+                    psiServices.GetComponent<IMenuModalUtil>(),
+                    psiServices.GetComponent<ICreateStepClassDialogUtil>(),
+                    psiServices.GetComponent<ICreateStepPartialClassFile>(),
+                    psiServices.GetComponent<SpecflowStepsDefinitionsCache>(),
+                    psiServices.GetComponent<ICreateSpecFlowStepUtil>()
+                ), BulbThemedIcons.RedBulb.Id, IntentionsAnchors.QuickFixesAnchor)
             };
         }